feat: update codebase to use modern go constructs (#1905)

+++ pkg/apis/defaults/jobscheduler.go
@@ -3,7 +3,7 @@ package defaults
 import "fmt"
 
 const RadixJobSchedulerPortName = "scheduler-port"
-const RadixJobTimeLimitSeconds = 43200 // 12 hours
+const RadixJobTimeLimitSeconds = int64(43200) // 12 hours
 
 // GetJobAuxKubeDeployName Get the aux kube deployment name for a job component
 func GetJobAuxKubeDeployName(jobName string) string {
